ucddrone/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-storages==1.14.2
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
oauth2client==4.1.3
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
